edit db

+++ entity/order.go
@@ -4,8 +4,8 @@ import "github.com/lib/pq"
 
 type Order struct {
 	Id             int            `json:"id"`
-	Weight         float32        `json:"weight" binding:"required"`
-	Price          float32        `json:"price" binding:"required"`
+	Weight         int            `json:"weight" binding:"required"`
+	Price          int            `json:"price" binding:"required"`
 	District       int            `json:"district" binding:"required"`
 	ConvenientTime pq.StringArray `json:"convenient_time" binding:"required"`
 }
